requirements.txt
aiohttp==3.10.6
colorama==0.4.4
cryptography==37.0.1
discord.py==2.5.2
qrcode==7.4.2
requests==2.24.0
websockets==10.3
